fixing apis

+++ tripGo/src/main/java/com/example/tripGo/repository/BookingRepository.java
@@ -3,5 +3,5 @@ package com.example.tripGo.repository;
 import com.example.tripGo.entity.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
+public interface BookingRepository extends JpaRepository<Booking, Long> {
 }
