+++ tripGo/src/main/java/com/example/tripGo/repository/BookingSeatRepository.java
@@ -3,5 +3,5 @@ package com.example.tripGo.repository;
 import com.example.tripGo.entity.BookingSeat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingSeatRepository extends JpaRepository<BookingSeat, Integer> {
+public interface BookingSeatRepository extends JpaRepository<BookingSeat, Long> {
 }
